requirements.txt
Flask==2.3.2
tensorflow
pillow
numpy
gtts
googletrans